chore: update to 1.21.132

+++ src-client/mc/client/gui/OperatorType.h
@@ -2,4 +2,4 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class OperatorType : int {};
+enum class OperatorType : uchar {};
